requirements.txt
yfinance==1.0
curl-cffi==0.7.3
pandas==2.1.4
python-dotenv==1.0.0
psycopg2-binary==2.9.9
requests==2.31.0
urllib3==2.1.0
flask==3.0.0
selenium==4.15.2
webdriver-manager==4.0.1